refactor: move proximity alerts logic and data to own cubit
DT-2256
