+++ lib/widgets/sliders/aircraft/detail/basic_fields.dart
@@ -3,6 +3,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_opendroneid/models/message_pack.dart';
 
 import '../../../../bloc/aircraft/aircraft_cubit.dart';
+import '../../../../bloc/proximity_alerts_cubit.dart';
 import '../../../../constants/colors.dart';
 import '../../../../utils/uasid_prefix_reader.dart';
 import '../../../../utils/utils.dart';
@@ -95,15 +96,18 @@ class BasicFields {
         children: [
           Text('This is my aircraft'),
           Checkbox(
-            value: context.read<AircraftCubit>().state.usersAircraftUASID !=
+            value: context
+                        .read<ProximityAlertsCubit>()
+                        .state
+                        .usersAircraftUASID !=
                     null &&
-                context.read<AircraftCubit>().state.usersAircraftUASID ==
+                context.read<ProximityAlertsCubit>().state.usersAircraftUASID ==
                     messagePackList.last.basicIdMessage?.uasId,
             onChanged: (value) {
               if (value != null &&
                   value &&
                   messagePackList.last.basicIdMessage?.uasId != null) {
-                context.read<AircraftCubit>().setUsersAircraftUASID(
+                context.read<ProximityAlertsCubit>().setUsersAircraftUASID(
                     messagePackList.last.basicIdMessage!.uasId);
               }
             },
